+++ jest-e2e.config.js
@@ -9,7 +9,7 @@ module.exports = {
     "^generator/(.*)$": "<rootDir>/src/generator/$1",
     "^e2e/(.*)$": "<rootDir>/src/e2e/$1",
     "^dependencies/(.*)$": "<rootDir>/src/dependencies/$1",
-    "^fileWriter/(.*)$": "<rootDir>/src/fileWriter/$1",
+    "^writer/(.*)$": "<rootDir>/src/writer/$1",
     "^parser/(.*)$": "<rootDir>/src/parser/$1",
     "^logger/(.*)$": "<rootDir>/src/logger/$1",
     "^utils/(.*)$": "<rootDir>/src/utils/$1",