+++ src/e2e/generate/prettier.test.ts
@@ -30,7 +30,14 @@ describe("Prettier eslint configuration Generator Mode", () => {
       },
     };
 
+    const tsConfig = {
+      compilerOptions: {
+        strict: true,
+      },
+    };
+
     testService.loadInputConfig(inputConfig);
+    testService.loadInitialTSConfig(tsConfig);
     await testService.runClinter();
     const outputConfig = testService.getParsedOutputConfig();
 
@@ -65,7 +72,15 @@ describe("Prettier eslint configuration Generator Mode", () => {
       },
     };
 
+    const tsConfig = {
+      compilerOptions: {
+        strict: true,
+      },
+    };
+
     testService.loadInputConfig(inputConfig);
+    testService.loadInitialTSConfig(tsConfig);
+
     await testService.runClinter();
     const outputConfig = testService.getParsedOutputConfig();
 
@@ -100,7 +115,15 @@ describe("Prettier eslint configuration Generator Mode", () => {
       },
     };
 
+    const tsConfig = {
+      compilerOptions: {
+        strict: true,
+      },
+    };
+
     testService.loadInputConfig(inputConfig);
+    testService.loadInitialTSConfig(tsConfig);
+
     await testService.runClinter();
     const outputConfig = testService.getParsedOutputConfig();