+++ src/e2e/generate/vue.test.ts
@@ -30,7 +30,14 @@ describe("Vue configuration Generator Mode", () => {
       },
     };
 
+    const tsConfig = {
+      compilerOptions: {
+        strict: true,
+      },
+    };
+
     testService.loadInputConfig(inputConfig);
+    testService.loadInitialTSConfig(tsConfig);
     await testService.runClinter();
     const outputConfig = testService.getParsedOutputConfig();