+++ src/generator/prettier.ts
@@ -1,14 +1,12 @@
+import { Linter } from "eslint";
+
 import { ESLintDependencyGenerator, ESLintGenerator } from "generator/types";
 import { identity, pipe } from "utils/utility";
 import { concatConfig, concatDependencies } from "generator/generate";
-import { FormatterInfo, FrontFrameworkInfo, TypescriptInfo } from "types";
-import {
-  prettierESLintConfig,
-  prettierESLintDependencies,
-  prettierReactESLintConfig,
-  prettierTypescriptESLintConfig,
-  prettierVueESLintConfig,
-} from "generator/base-configs";
+import { FormatterInfo, TypescriptInfo } from "types";
+import { prettierESLintConfig, prettierESLintDependencies, tsOverrider } from "generator/base-configs";
+import { wrapConfigInOverride } from "generator/override";
+import { cleanESLintExtendsField } from "generator/generate";
 
 const generatePrettierTypescriptESLintConfig: ESLintGenerator = (userAnswers) => {
   switch (userAnswers.typescript) {
@@ -17,20 +15,7 @@ const generatePrettierTypescriptESLintConfig: ESLintGenerator = (userAnswers) =>
 
     case TypescriptInfo.NoTypeChecking:
     case TypescriptInfo.WithTypeChecking:
-      return concatConfig(prettierTypescriptESLintConfig);
-  }
-};
-
-const generatePrettierReactESLintConfig: ESLintGenerator = (userAnswers) => {
-  switch (userAnswers.frontFramework) {
-    case FrontFrameworkInfo.React:
-      return concatConfig(prettierReactESLintConfig);
-
-    case FrontFrameworkInfo.Vue:
-      return concatConfig(prettierVueESLintConfig);
-
-    default:
-      return identity;
+      return concatConfig(wrapConfigInOverride(tsOverrider)(prettierESLintConfig));
   }
 };
 
@@ -40,11 +25,7 @@ export const generatePrettierESlintConfig: ESLintGenerator = (userAnswers) => {
       return identity;
 
     case FormatterInfo.Prettier:
-      return pipe(
-        concatConfig(prettierESLintConfig),
-        generatePrettierTypescriptESLintConfig(userAnswers),
-        generatePrettierReactESLintConfig(userAnswers)
-      );
+      return pipe(concatConfig(prettierESLintConfig), generatePrettierTypescriptESLintConfig(userAnswers));
   }
 };
 
@@ -56,4 +37,8 @@ export const getPrettierESLintDependencies: ESLintDependencyGenerator = (userAns
     case FormatterInfo.Prettier:
       return concatDependencies(prettierESLintDependencies);
   }
+};
+
+export const cleanPrettierConfig = (config: Linter.Config): Linter.Config => {
+  return cleanESLintExtendsField(config, (item: string) => !item.startsWith("prettier/"));
 };